+++ bot/Dockerfile
@@ -26,4 +26,4 @@ RUN pipenv install --system --deploy
 # Expose port for debugging
 #EXPOSE 5678:5678
 
-CMD ["python", "main.py"]
+CMD ./wait_for_db.sh mysql:3306 -- python main.py
