+++ bot/mods/user_chat_time_logging.py
@@ -1,14 +1,15 @@
 from datetime import datetime
 from typing import Union
 
-from mods.database_models import Users
 from twitchbot import cfg
 from twitchbot import Channel
 from twitchbot import channels
 from twitchbot import get_user_id
 from twitchbot import Message
 from twitchbot import Mod
-from twitchbot.database.session import session
+
+from mods._database_models import session
+from mods._database_models import Users
 
 
 class UserChatTimeLogging(Mod):
